BUILD: fix compilation problem on clang-16 with hip
The upcoming rocm release switched from clang-15 to clang-16,
and we observer a compilation failure of the ec/rocm/kernel code.
This commit fixes the compilation problem by omitting the
`-Xcompile` flag in cuda_lt.sh if the compiler is hipcc.
(Previously, the flag generated a warning, now its an error).

Signed-off-by: Edgar Gabriel <Edgar.Gabriel@amd.com>